refactor(console): move dialogs to specific folder + clean up dialog buttons
