+++ gravitee-apim-console-webui/src/management/api/portal/ng-subscriptions/components/dialogs/change-end-date/api-portal-subscription-change-end-date-dialog.component.html
@@ -1,13 +1,13 @@
 <!--
 
     Copyright (C) 2015 The Gravitee team (http://gravitee.io)
-    
+
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
     You may obtain a copy of the License at
-    
+
             http://www.apache.org/licenses/LICENSE-2.0
-    
+
     Unless required by applicable law or agreed to in writing, software
     distributed under the License is distributed on an "AS IS" BASIS,
     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,15 +15,19 @@
     limitations under the License.
 
 -->
-<form [formGroup]="form" *ngIf="form">
-  <h2 mat-dialog-title class="title">Change your API Key's expiration date</h2>
+<form [formGroup]="form" *ngIf="form" (ngSubmit)="onClose()">
+  <h2 mat-dialog-title class="title">Change the subscription end date</h2>
 
-  <mat-dialog-content class="expire-api-key__dialog-content">
-    <div class="expire-api-key__content">
+  <mat-dialog-content class="change-end-date__dialog-content">
+    <div class="change-end-date__content">
+      <div class="change-end-date__warning">
+        Be careful, by changing the end date, {{ data.applicationName }} may no longer have access to this API.
+      </div>
+      <div *ngIf="data.securityType === 'API_KEY'">All API keys after the new end date will be revoked.</div>
       <mat-form-field appearance="outline">
         <mat-label>Choose a date</mat-label>
-        <div class="expire-api-key__datepicker">
-          <input matInput [min]="minDate" [matDatepicker]="picker" formControlName="expirationDate" />
+        <div class="change-end-date__datepicker">
+          <input matInput [min]="minDate" [matDatepicker]="picker" formControlName="endDate" />
           <mat-datepicker-toggle matIconSuffix [for]="picker"></mat-datepicker-toggle>
           <mat-datepicker #picker></mat-datepicker>
         </div>
@@ -32,16 +36,15 @@
     </div>
   </mat-dialog-content>
   <mat-dialog-actions class="actions">
-    <button mat-flat-button [mat-dialog-close]="undefined">Cancel</button>
+    <button mat-flat-button type="button" [mat-dialog-close]="undefined">Cancel</button>
     <button
       color="primary"
       type="submit"
       mat-raised-button
-      aria-label="Change expiration date of API Key"
-      (click)="onClose()"
+      aria-label="Change end date of subscription"
       [disabled]="form.invalid || form.pristine"
     >
-      Change expiration date
+      Change end date
     </button>
   </mat-dialog-actions>
 </form>
