+++ gravitee-apim-console-webui/src/management/api/portal/ng-subscriptions/components/dialogs/change-end-date/api-portal-subscription-change-end-date-dialog.component.scss
@@ -1,5 +1,5 @@
-@use '@angular/material' as mat;
-@use '@gravitee/ui-particles-angular' as gio;
+@use 'node_modules/@angular/material/index' as mat;
+@use 'node_modules/@gravitee/ui-particles-angular/index' as gio;
 
 .change-end-date {
   &__content {
