+++ gravitee-apim-console-webui/src/management/api/portal/ng-subscriptions/components/dialogs/expire-api-key/api-portal-subscription-expire-api-key-dialog.component.scss
