+++ gravitee-apim-console-webui/src/management/api/portal/ng-subscriptions/components/dialogs/reject/api-portal-subscription-reject-dialog.component.scss
