+++ gravitee-apim-console-webui/src/management/api/portal/ng-subscriptions/components/dialogs/renew/api-portal-subscription-renew-dialog.component.html
@@ -1,13 +1,13 @@
 <!--
 
     Copyright (C) 2015 The Gravitee team (http://gravitee.io)
-    
+
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
     You may obtain a copy of the License at
-    
+
             http://www.apache.org/licenses/LICENSE-2.0
-    
+
     Unless required by applicable law or agreed to in writing, software
     distributed under the License is distributed on an "AS IS" BASIS,
     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,28 +15,17 @@
     limitations under the License.
 
 -->
-<h2 matDialogTitle>Reject your subscription</h2>
-
-<form [formGroup]="form">
-  <mat-dialog-content class="reject-subscription__content">
-    <div class="mat-body-1">Explain the reason for the rejection</div>
-    <div class="reject-subscription__reason">
-      <mat-form-field appearance="outline" class="reject-subscription__reason__input">
-        <input matInput formControlName="reason" />
-      </mat-form-field>
+<h2 matDialogTitle>Renew your API Key</h2>
+<form [formGroup]="form" (ngSubmit)="onClose()">
+  <mat-dialog-content class="renew__content">
+    <div class="mat-body-1">Your previous API Key will be no longer valid in 2 hours!</div>
+    <div *ngIf="data.canUseCustomApiKey" class="renew__content__custom-api-key">
+      <div class="mat-body-1">You can provide a custom API key here</div>
+      <api-key-validation formControlName="apiKey" [applicationId]="data.applicationId" [apiId]="data.apiId"></api-key-validation>
     </div>
   </mat-dialog-content>
   <mat-dialog-actions class="actions">
-    <button mat-flat-button [mat-dialog-close]="undefined">Cancel</button>
-    <button
-      color="primary"
-      type="submit"
-      mat-raised-button
-      aria-label="Reject the subscription"
-      (click)="onClose()"
-      [disabled]="form.invalid"
-    >
-      Reject
-    </button>
+    <button mat-flat-button type="button" [mat-dialog-close]="undefined">Cancel</button>
+    <button color="primary" type="submit" mat-raised-button aria-label="Renew API Key" [disabled]="form.invalid">Renew</button>
   </mat-dialog-actions>
 </form>
