+++ gravitee-apim-console-webui/src/management/api/portal/ng-subscriptions/components/dialogs/transfer/api-portal-subscription-transfer-dialog.component.html
@@ -15,7 +15,7 @@
     limitations under the License.
 
 -->
-<form [formGroup]="form" *ngIf="form">
+<form [formGroup]="form" *ngIf="form" (ngSubmit)="onClose()">
   <h2 mat-dialog-title class="title">Transfer your subscription</h2>
 
   <mat-dialog-content class="subscription-transfer__dialog-content">
@@ -36,14 +36,13 @@
     </div>
   </mat-dialog-content>
   <mat-dialog-actions class="actions">
-    <button mat-flat-button [mat-dialog-close]="undefined">Cancel</button>
+    <button mat-flat-button type="button" [mat-dialog-close]="undefined">Cancel</button>
     <button
       *ngIf="plans?.length > 0"
       color="primary"
       type="submit"
       mat-raised-button
       aria-label="Transfer subscription"
-      (click)="onClose()"
       [disabled]="form.invalid"
     >
       Transfer
